ppt_remote/requirements_exe.txt
pyinstaller==6.3.0
pyautogui==0.9.54
Pillow==11.2.1
keyboard==0.13.5 